+++ leda/components/Tooltip/TooltipBody.tsx
@@ -10,6 +10,9 @@ export const TooltipBody = React.forwardRef((props: TooltipBodyProps, ref?: Reac
     title,
   } = props;
 
+  // Width should be set explicitly for correct displaying in IE
+  const { width } = style;
+
   const tooltip = (
     <div
       className={className}
@@ -17,7 +20,7 @@ export const TooltipBody = React.forwardRef((props: TooltipBodyProps, ref?: Reac
       style={style}
       onTransitionEnd={handleTransitionEnd}
     >
-      <div>
+      <div style={{ width }}>
         {title}
       </div>
     </div>